@@ -6,18 +6,26 @@
     "pnpm": ">=3"
   },
   "dependencies": {
+    "@c3/types": "workspace:^1.0.4",
+    "@scriptbot/cli": "^1.0.150",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "antd": "^4.18.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "msw": "^0.47.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "resolutions": {
-    "@types/react": "17.0.2",
-    "@types/react-dom": "17.0.2"
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.2",
     "@babel/plugin-transform-modules-commonjs": "^7.17.7",
-    "@c3/cli": "^1.0.105",
+    "@babel/preset-env": "^7.18.10",
+    "@babel/preset-typescript": "^7.18.6",
     "@changesets/cli": "^2.22.0",
     "@jest/types": "^28.1.1",
     "@pnpm/sort-packages": "^3.0.7",
@@ -31,13 +39,14 @@
     "@storybook/manager-webpack4": "^6.5.9",
     "@storybook/react": "^6.5.9",
     "@storybook/testing-library": "^0.0.13",
-    "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.178",
+    "@types/jest": "^27.5.2",
+    "@types/lodash": "^4.14.182",
     "@types/node": "^17.0.13",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.15.0",
     "@typescript-eslint/parser": "^5.15.0",
+    "babel-jest": "^29.0.1",
     "babel-loader": "^8.2.3",
     "eslint": "^8.11.0",
     "eslint-plugin-import": "^2.25.4",
@@ -47,11 +56,14 @@
     "find-packages": "^9.0.9",
     "husky": ">=6",
     "jest": "^27.5.1",
+    "jest-environment-jsdom": "^29.0.1",
     "jest-extended": "^2.0.0",
     "js-beautify": "^1.14.0",
     "lint-staged": ">=10",
+    "lodash": "^4.17.21",
     "pkgs-graph": "^7.0.2",
     "prettier": "2.6.1",
+    "semver": "^7.3.7",
     "tslib": "^2.3.1",
     "typescript": "^4.5.5",
     "vite": "^2.9.12",
@@ -64,7 +76,8 @@
     "ci:publish": "zx scripts/publish.mjs publish",
     "clean": "rm -rf dist && rm -rf node_modules ",
     "reset": "pnpm clean && pnpm install",
-    "storybook": "start-storybook -p 6007"
+    "storybook": "start-storybook -p 6007",
+    "test": "jest"
   },
   "lint-staged": {
     "*.{ts,tsx}": [],
