@@ -1,6 +1,5 @@
-import { useMount, useSwitch } from '@c3/hooks';
 import { IAPI, RawReqParameter, RawResBody, ReqParameter, ResBody } from './makeApi/api';
-import { useCallback, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
 export type UseApiOption<T extends RawReqParameter> = {
   fetchOnMounted: boolean;
@@ -13,30 +12,32 @@ export const useApi = <
   _ResBody extends ResBody
 >(
   api: IAPI<_RawReqParameter, _ReqParameter, _RawResBody, _ResBody>,
-  option: UseApiOption<_RawReqParameter> = { fetchOnMounted: true }
+  option: UseApiOption<_RawReqParameter> = { fetchOnMounted: false }
 ) => {
   const [data, setData] = useState(api.defaultData);
-  const [loading, on, off] = useSwitch(false);
+  const [loading, setLoading] = useState(false);
   const fetch = useCallback(
-    async (rrp?: _RawReqParameter) => {
+    async (rrp?: _RawReqParameter, ...args: any[]) => {
       try {
-        on();
-        const res = await api.fetch(rrp);
+        setLoading(true);
+        const res = await api.fetch(rrp, ...args);
         if (!res) {
           setData(api.defaultData);
+          return api.defaultData;
         }
         setData(res);
+        return res;
       } finally {
-        off();
+        setLoading(false);
       }
     },
-    [off, on, api]
+    [api]
   );
-  useMount(() => {
+  useEffect(() => {
     if (option.fetchOnMounted) {
       fetch(option.defaultReqParameter);
     }
-  });
+  }, [fetch, option.defaultReqParameter, option.fetchOnMounted]);
 
   return [data, fetch, setData, loading] as const;
 };
