@@ -1,11 +1,6 @@
-import { initCryptoDebug } from './utils';
-import { useOnChainChange, useWallet_ } from './wallet';
-
+export * from './wallet';
 export * from './context';
 export * from './contract';
 export * from './network';
-export * from './provider';
 export * from './utils';
-export { useWallet_, useOnChainChange };
 
-initCryptoDebug();
