@@ -1,3 +1,2 @@
-export * from './chain';
-export * from './chainIds';
+export * from './types';
 export * from './chains';
