@@ -1,4 +1,6 @@
-export * from './useAccount';
-export * from './useBalance';
+export * from './useAccount_';
+export * from './useWallet_';
+export * from './getWalletAccount';
+export * from './useContractViaWallet';
 export * from './useOnChainChange';
-export * from './useWallet';
+export * from './walletProvider';
