@@ -1,4 +1,5 @@
-import { assert, ID, IDable, isEqual, isNullish } from '@c3/utils';
+import { assert, isEqual, isNullish } from '@c3/utils';
+import { ID, IDable } from '@c3/types';
 import { useCallback } from 'react';
 
 type SwitchOnFn = (id: ID) => void;
