@@ -12,6 +12,6 @@ export const mock = {
   getId: () => `uniq-id-${gid++}`,
   getRandomUser,
   getRandomColor: () => {
-    return '#' + Math.floor(Math.random() * 16777215).toString(16);
+    return `#${Math.floor(Math.random() * 16777215).toString(16)}`;
   },
 };
