@@ -1,8 +1,8 @@
-import { Color, PercentString } from '../lang/famous';
+import { PercentString } from '@c3/types';
 
 export type GradientStop = {
   offset: number | PercentString;
-  stopColor: Color;
+  stopColor: string;
   stopOpacity?: number; //0-1
 };
 export type LinearGradient = GradientStop[];
