@@ -1,4 +1,6 @@
-import { Fn } from './famous';
+import { Fn } from '@c3/types';
 
-export const pipe = (...fns: Fn[]) => (...args: any[]) =>
-  fns.reduceRight((res, fn) => [fn.call(null, ...res)], args)[0];
+export const pipe =
+  (...fns: Fn[]) =>
+  (...args: any[]) =>
+    fns.reduceRight((res, fn) => [fn.call(null, ...res)], args)[0];
